fr-google-jukebox-musicgen/requirements.txt
uvicorn==0.32.1
Wave==0.0.2
google-cloud-storage==2.19.0
fastapi==0.115.6
httpx==0.28.1
pydantic==2.10.3
transformers==4.47.0
python-dotenv==1.0.1